% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module.R
\name{combView}
\alias{combView}
\title{shiny module server - view curves}
\usage{
combView(input, output, session, deg, syn, deg.ratio, syn.ratio, deg.t, syn.t,
  tcc)
}
\arguments{
\item{input}{shiny input}

\item{output}{shiny output}

\item{session}{shiny session}

\item{deg}{a reactive object, containing an object returned by fitNLSModels}

\item{syn}{a reactive object, containing an object returned by fitNLSModels}

\item{deg.ratio}{a reactive object, a matrix of degradation ratios at differnet time points}

\item{syn.ratio}{a reactive object, a matrix of synthesis ratios at differnet time points}

\item{deg.t}{a reactive object, times points for degradation ratios}

\item{syn.t}{a reactive object, time points for synthesis ratios}

\item{tcc}{a reactive object, cell doubling time}
}
\value{
write table to disk
}
\description{
Should not be called by users. shiny module server - view curves
}
\keyword{internal}
