% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module.R
\name{combViewUI}
\alias{combViewUI}
\title{shiny module, view UI}
\usage{
combViewUI(id)
}
\arguments{
\item{id}{id for namespace}
}
\value{
no value to be returned
}
\description{
Should not be called by users. shiny module, view UI
}
\keyword{internal}
