% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionsSimPlot.R
\name{degCurve}
\alias{degCurve}
\alias{synCurve}
\title{Calculating ratio of proteins turnover using theoretical formula}
\usage{
degCurve(A, B, kd, tcc, t)

synCurve(A, B, kd, tcc, t)
}
\arguments{
\item{A}{model parameter A (amplitude), see details}

\item{B}{model parameter B (offset), see details.}

\item{kd}{model parameter kd (degradation/synthesis constant), see details.}

\item{tcc}{model parameter tcc (doubling time), see details.}

\item{t}{The time points (hours)}
}
\value{
A vector of ratios of protein turned over
}
\description{
Generate ratio of proteins turnover using theoretical formula. This used for control/unit test.
}
\details{
Generate
 withMathJax("$$f(t)=(A-B)\\cdot e^{-(k_d+\\frac{ln2}{t_{cc}})\\cdot t}+B$$")
 "The model fitted for the synthesis curve is as:"
 withMathJax("$$f(t)=(B-A)\\cdot e^{-(k_s+\\frac{ln2}{t_{cc}})\\cdot t}+A$$")
 "where t is the time points (given in hours)
}
\examples{
 times <- 2^(0:6)
 drat <- degCurve(A = 1, B = 0, kd = 0.05, tcc = Inf, times)
 plot(times, drat)
 times <- 2^(0:6)
 srat <- synCurve(A = 1, B = 0, kd = 0.05, tcc = Inf, times)
 plot(times, srat)
}
\keyword{internal}
