% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionsSimPlot.R
\name{ebar}
\alias{ebar}
\title{draw horizontal error bars}
\usage{
ebar(x, y, lwd = 2, col = 1, lty = 1, length = 0.1, pch = 18)
}
\arguments{
\item{x}{a vector of length 2}

\item{y}{a single value or a numerical vector specify the position on y-axis}

\item{lwd}{line width}

\item{col}{color}

\item{lty}{the lty}

\item{length}{the length of vertical bars at the end of arrow bars}

\item{pch}{shape for points of mean value}
}
\value{
no value to be returned
}
\description{
draw horizontal error bar, not called by users
}
\examples{
# plot(0, 0)
# ebar(c(-0.1, 0.1), c(0), lty = 2, lwd = 2, col = "red")
# ebar(c(-0.1, 0.1), c(0.1, 0, -0.4), lty = 2)

}
