% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module.R
\name{fmod}
\alias{fmod}
\title{shiny module - fitting in server}
\usage{
fmod(input, output, session, x, tcc = reactive(Inf), f, time, type, A, B,
  par.init, par.lower, par.upper, pre.col, ncore, resultPath)
}
\arguments{
\item{input}{shiny input}

\item{output}{shiny output}

\item{session}{shiny session}

\item{x}{a reactive object, matrix, passed to fitNLSModels}

\item{tcc}{a reactive object, cell doubling time}

\item{f}{a reactive object, f passed to fitNLSModels}

\item{time}{a reactive object, time points passed to fitNLSModels}

\item{type}{a reactive object, the type of fitting (syn or deg) passed to fitNLSModels}

\item{A}{a reactive object, passed to fitNLSModels}

\item{B}{a reactive object, passed to fitNLSModels}

\item{par.init}{a reactive object, passed to fitNLSModels}

\item{par.lower}{a reactive object, passed to fitNLSModels}

\item{par.upper}{a reactive object, passed to fitNLSModels}

\item{pre.col}{a reactive object containing a data.frame, for the annotation columns}

\item{ncore}{a reactive object, passed to fitNLSModels}

\item{resultPath}{where the figures to be saved}
}
\value{
a reactive value
  reactive( list(pre.col = pre.col(), mat = r()$mat, list = r()$list, type = r()$type) )
}
\description{
Should not be called by users. shiny module for server, curve fitting
}
\keyword{internal}
