% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module.R
\name{fmodUI}
\alias{fmodUI}
\title{shiny module, curve fitting UI}
\usage{
fmodUI(id)
}
\arguments{
\item{id}{id for namespace}
}
\value{
no value to be returned
}
\description{
Should not be called by users. shiny module, curve fitting UI
}
\keyword{internal}
