% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionsSimPlot.R
\name{plotCurve.one}
\alias{plotCurve.one}
\title{plot a single fit of degradation or synthesis curve}
\usage{
plotCurve.one(x, t, tcc = Inf, col = "red", curve = NULL, pch = 20,
  lty = 1, err.lwd = 2, main = "")
}
\arguments{
\item{x}{an object returned by \code{\link{fitSynNLS}} or \code{\link{fitDegNLS}} with vector input}

\item{t}{time points}

\item{tcc}{cell doubling time}

\item{col}{color, passed to plot}

\item{curve}{what type of curve fit, if is not null, should be either "degradation" or "synthesis"}

\item{pch}{passed to plot}

\item{lty}{passed to plot}

\item{err.lwd}{the line width of error bars}

\item{main}{main title, passed to plot}
}
\value{
a plot generated, returns nothing
}
\description{
plot a single fit of degradation or synthesis curve
}
\note{
This function won't work if input of \code{\link{fitSynNLS}} or \code{\link{fitDegNLS}} is a one row matrix
}
\examples{
  tp <- c(0, 1, 2, 4, 8, 16, 32, 64)
  ratios <- degCurve(A=0.85, B = 0.1, kd=0.5, tcc=Inf, t = tp) + rnorm(length(tp), sd = 0.05)
  # vector input
  r <- fitDegNLS(ratios, t = tp, tcc = Inf)
  plotCurve.one(r, tp)
}
