% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionsSimPlot.R
\name{plotCurve}
\alias{plotCurve}
\title{plot protein degration/synthesis curve}
\usage{
plotCurve(x, t, tcc, A, B, k, add = TRUE, col = "red", lineOnly = FALSE,
  curve = c("degradation", "synthesis")[1], pch = 20, lty = 1,
  ylim = NULL, err.x = NULL, err.y = NULL, err.lwd = 2, main = "")
}
\arguments{
\item{x}{The data point, ratio, if \code{lineOnly=TRUE}, this argument is ignored}

\item{t}{time points (hours)}

\item{tcc}{doubling time of cell lines}

\item{A}{model parameter A (amplitude), see details}

\item{B}{model parameter B (offset), see details.}

\item{k}{model parameter kd (degradation constant) or sd (synthesis constant), see details.}

\item{add}{if the plot should be added on top of another plot}

\item{col}{The color of line}

\item{lineOnly}{logical, whether only draw the fitted line, ie. the input data point is ignored}

\item{curve}{which curve want to draw, should be either "degradation" or "synthesis"}

\item{pch}{passed to \code{plot}}

\item{lty}{passed to \code{line}}

\item{ylim}{passed to \code{plot}}

\item{err.x}{draw error bar}

\item{err.y}{draw error bar}

\item{err.lwd}{the line width of error bars}

\item{main}{the main title of plot}
}
\value{
no value to be returned
}
\description{
plot protein degration/synthesis curve with origin data point and fitted curve
}
\examples{
 times <- 2^(0:6)
 drat <- degCurve(A = 1, B = 0, kd = 0.05, tcc = Inf, times)
 drat <- drat + rnorm(length(times), sd = 0.2)
 plot(times, drat)
 plotCurve(drat, t = times, tcc = Inf, A = 1, B = 0, k = 0.05, add=FALSE, col="red",
       lineOnly = FALSE, curve = "degradation", ylim = NULL,
       err.x = c(11, 20), err.y = 0.5, main = "toy curve fitting")
}
