% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runShiny.R
\name{runShiny}
\alias{runShiny}
\title{Lanch shiny application for protein degradation/synthesis curve fitting}
\usage{
runShiny(maxFileSize = 100 * 1024^2, figureFolder = getwd())
}
\arguments{
\item{maxFileSize}{max size of files can be uploaded}

\item{figureFolder}{a character string to specify a directory where the figures to be saved}
}
\value{
A shiny application will be launched, not values to be returned
}
\description{
Lanch shiny application for protein degradation/synthesis curve fitting
}
\examples{
# not run
1
# runShiny()
}
