% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refitwoOutlier.R
\name{sigDF}
\alias{sigDF}
\title{rounds the values in a data.frame}
\usage{
sigDF(x, digits = 3)
}
\arguments{
\item{x}{a \code{data.frame} or \code{list}}

\item{digits}{teh number of significant digits}
}
\value{
the same object and size as x, but all numerical values are rounded
}
\description{
for a data.frame or list, round all numerical variables to specified number of significant digits
}
\examples{
a <- data.frame(num = rnorm(4),
                char = letters[1:4])
sigDF(a, 3)
}
\keyword{internal}
