% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapseIntensity.R
\name{collapseIntesity}
\alias{collapseIntesity}
\title{Collapse intensity}
\usage{
collapseIntesity(x, intensity.col, collapse.col = "Sequence", annot.col,
  n.cores = 1, collapse.method = c("sum", "mean", "median", "max.range",
  "max.range01")[1])
}
\arguments{
\item{x}{a \code{data.frame} containing at least the intensity columns and a column according to which
intensities are collapsed}

\item{intensity.col}{the column of intensities, could be either column name or column index}

\item{collapse.col}{the column accroding to which intensity are collapsed}

\item{annot.col}{the annotation columns, which will be retained in the output}

\item{n.cores}{the number of cores to be used, it is passed to mclapply function}

\item{collapse.method}{the collapse method, see \code{details} section.}
}
\value{
a \code{data.frame} of collapsed intensities, containing the collapsed.col, intensity.col and annot.cols.
}
\description{
Collpase intensity to peptide or protein level
}
\details{
This function prepare input for \code{\link{fitDegNLS}}. Two typical usages are: 2) collapse identical
  peptides, here the peptide sequence should be used as collapse column; 2) collapse from peptide level to protein
  levels, so the protein name or ID could be used as collapse column. The following collapse methods are avaliable:
  \itemize{
    \item sum sum the intensities;
    \item mean mean of intensities;
    \item median median of intensities;
    \item max.range retain the one with max range;
    \item max.range01 retain the one with max difference between first and last intensity columns.
  }
}
\examples{
  df <- data.frame(
    sequence = c("CCAD", "CSDS", "IIURE", "IIURE", "XMNT"),
    t0 = rnorm(5),
    t1 = rnorm(5),
    t2 = rnorm(5),
    name = LETTERS[1:5]
  )
  collapseIntesity(x = df, collapse.col = "sequence", intensity.col = c("t0", "t1", "t2"), annot.col = "name")
  collapseIntesity(x = df, collapse.col = 1, intensity.col = 2:4, annot.col = 5, collapse.method = "max.range01")
}
\author{
Chen Meng
}
